feat: temporarily allow all actions from components (#21)
This change adds certificate chain validation for authenticating clients
using certificates issued by the certificate manager. If certificate
validation fails, then the code falls back to the previous mechanism (by
checking with IoT Core to see if the certificate belongs to a Thing).

A special session ID is returned which gives these clients permissions
to perform any action. This is a temporary workaround, as we'd like to
be able to restrict which actions a component may perform.