fix: dropped CIS updates in CertificateManager (#44)
This change fixes a race condition in the CertificateManager
subscribeToServerCertificateUpdates method. If a CIS update occurs while
the initial server certificate is being generated, but before the
generator is added to the shadow monitor, then the CIS update event will
be missed. This will result in the server certificate not including
correct connectivity information.