refactor: create session using attribute providers
This change also removes interface methods that result in updating
sessions, which results in sessions becoming immutable by contract
instead of by convention (technically the underlying attribute providers
can still change so this isn't 100% true)

It's TBD on whether this is actually the behavior we want. But for now
this allows us to delete unused code.